Fix runner doctor audit findings
